+++ packages/@expo/cli/src/export/exportStaticAsync.ts
@@ -30,6 +30,7 @@ import { getFreePortAsync } from '../utils/port';
 const debug = require('debug')('expo:export:generateStaticRoutes') as typeof console.log;
 
 type Options = {
+  mode: 'production' | 'development';
   files?: ExportAssetMap;
   outputDir: string;
   minify: boolean;
@@ -38,9 +39,9 @@ type Options = {
   includeSourceMaps: boolean;
   entryPoint?: string;
   clear: boolean;
-  asyncRoutes: boolean;
   routerRoot: string;
   maxWorkers?: number;
+  isExporting: boolean;
 };
 
 type HtmlRequestLocation = {
@@ -65,19 +66,23 @@ export async function unstable_exportStaticAsync(projectRoot: string, options: O
   // TODO: Prevent starting the watcher.
   const devServerManager = new DevServerManager(projectRoot, {
     minify: options.minify,
-    mode: 'production',
+    mode: options.mode,
     port,
+    isExporting: true,
     location: {},
     resetDevServer: options.clear,
     maxWorkers: options.maxWorkers,
   });
+
   await devServerManager.startAsync([
     {
       type: 'metro',
       options: {
         port,
+        mode: options.mode,
         location: {},
         isExporting: true,
+        minify: options.minify,
         resetDevServer: options.clear,
         maxWorkers: options.maxWorkers,
       },
@@ -173,17 +178,10 @@ function makeRuntimeEntryPointsAbsolute(manifest: ExpoRouterRuntimeManifest, app
 async function exportFromServerAsync(
   projectRoot: string,
   devServerManager: DevServerManager,
-  {
-    outputDir,
-    baseUrl,
-    exportServer,
-    minify,
-    includeSourceMaps,
-    routerRoot,
-    asyncRoutes,
-    files = new Map(),
-  }: Options
+  { outputDir, baseUrl, exportServer, includeSourceMaps, routerRoot, files = new Map() }: Options
 ): Promise<ExportAssetMap> {
+  const platform = 'web';
+  const isExporting = true;
   const appDir = path.join(projectRoot, routerRoot);
   const injectFaviconTag = await getVirtualFaviconAssetsAsync(projectRoot, {
     outputDir,
@@ -196,20 +194,9 @@ async function exportFromServerAsync(
 
   const [resources, { manifest, serverManifest, renderAsync }] = await Promise.all([
     devServer.getStaticResourcesAsync({
-      isExporting: true,
-      mode: 'production',
-      minify,
       includeSourceMaps,
-      baseUrl,
-      asyncRoutes,
-      routerRoot,
     }),
-    devServer.getStaticRenderFunctionAsync({
-      mode: 'production',
-      minify,
-      baseUrl,
-      routerRoot,
-    }),
+    devServer.getStaticRenderFunctionAsync(),
   ]);
 
   makeRuntimeEntryPointsAbsolute(manifest, appDir);
@@ -223,7 +210,7 @@ async function exportFromServerAsync(
     async renderAsync({ pathname, route }) {
       const template = await renderAsync(pathname);
       let html = await serializeHtmlWithAssets({
-        mode: 'production',
+        isExporting,
         resources: resources.artifacts,
         template,
         baseUrl,
@@ -239,7 +226,7 @@ async function exportFromServerAsync(
   });
 
   getFilesFromSerialAssets(resources.artifacts, {
-    platform: 'web',
+    platform,
     includeSourceMaps,
     files,
   });
@@ -249,7 +236,7 @@ async function exportFromServerAsync(
     // NOTE(kitten): Re. above, this is now using `files` except for iOS catalog output, which isn't used here
     await persistMetroAssetsAsync(resources.assets, {
       files,
-      platform: 'web',
+      platform,
       outputDirectory: outputDir,
       baseUrl,
     });
@@ -259,9 +246,7 @@ async function exportFromServerAsync(
     const apiRoutes = await exportApiRoutesAsync({
       outputDir,
       server: devServer,
-      routerRoot,
       manifest: serverManifest,
-      baseUrl,
     });
 
     // Add the api routes to the files to export.
@@ -422,22 +407,14 @@ export function getPathVariations(routePath: string): string[] {
 async function exportApiRoutesAsync({
   outputDir,
   server,
-  routerRoot,
-  baseUrl,
   ...props
-}: {
-  outputDir: string;
+}: Pick<Options, 'outputDir'> & {
   server: MetroBundlerDevServer;
-  routerRoot: string;
   manifest: ExpoRouterServerManifestV1;
-  baseUrl: string;
 }): Promise<ExportAssetMap> {
   const { manifest, files } = await server.exportExpoRouterApiRoutesAsync({
-    mode: 'production',
-    routerRoot,
     outputDir: '_expo/functions',
     prerenderManifest: props.manifest,
-    baseUrl,
   });
 
   Log.log(chalk.bold`Exporting ${files.size} API Routes.`);