+++ packages/@expo/cli/src/start/server/metro/metroErrorInterface.ts
@@ -256,11 +256,12 @@ export async function getErrorOverlayHtmlAsync({
     '',
     resolveFrom(projectRoot, 'expo-router/_error'),
     {
-      dev: true,
+      mode: 'development',
       platform: 'web',
       minify: false,
       baseUrl: '',
       routerRoot,
+      isExporting: false,
     }
   );